requirements.txt
appwrite==9.0.2
asgiref==3.8.1
bcrypt==4.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==4.1
django-showurls==1.0.1
djangorestframework==3.15.2
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
msgpack==1.1.0
nltk==3.9.1
numpy==2.1.3
packaging==25.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyngrok==7.2.8
pyparsing==3.2.1
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
